restart method in start.java
start is now namespaced to the run package to allow importing
from GameWindow.

start.restart() flushes the EventQueue and completely resets the
application. This is a weird solution to restart the game but
it's the only way I could figure it without knowing a lot more
about how EventQueue and Swing threads function.

Search playstyles now also correctly report the number of
nodes searched. There's a lot!